npm scripts to handle moving static files
